% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agreement.R
\name{agreement}
\alias{agreement}
\title{Agreement}
\usage{
agreement(a, b = NULL)
}
\arguments{
\item{x}{A numerical vector or a matrix of 2 raters}

\item{y}{A numerical vector}
}
\value{
overall agreement, specific agreements for each class and the confusion matrix
}
\description{
This function returns the overall and specific agreeement (Cicchetti 1990)
}
\examples{
agreement(c(1,2,2,3,2), c(1,2,4,3,2))
}
