% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAanalysis.R
\name{BA.analysis}
\alias{BA.analysis}
\title{Bland-Altman analysis}
\usage{
BA.analysis(x, y, ...)
}
\arguments{
\item{x}{A numerical vector}

\item{y}{A numerical vector}

\item{perrcent}{bool for relative estimation.}
}
\value{
All the element required to a BA analysis
}
\description{
This function returns the Bland-Altman elements with 95% confidence
interval on the bias and on the limits of agreements. It also 
calculates the percentage of error. Data can be converted to 
relative estimation.
}
\details{
References: 
1) Bland, J.M. & Altman, D.G., 1986. Statistical methods for assessing agreement
between two methods of clinical measurement. Lancet, 1(8476), pp.307–310.
2) Giavarina, D., 2015. Understanding Bland Altman analysis. Biochemia Medica, 25(2), pp.141–151
}
\examples{
BA.analysis(c(1,2,3,4), c(2,3,4,5))
}
