% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAplot.R
\name{BA.plot}
\alias{BA.plot}
\title{Bland-Altman plot}
\usage{
BA.plot(x, y, ...)
}
\arguments{
\item{x}{and y: numerical vector, title of the plot and percent (bool) for relative
estimation.}
}
\value{
standard R plot
}
\description{
This function creates a Bland-Altman plot with 95% confidence
interval on the bias and on the limits of agreements. It also 
calculates the percentage of error. Data can be converted to 
relative estimation.
}
\details{
References: 
1) Bland, J.M. & Altman, D.G., 1986. Statistical methods for assessing agreement
between two methods of clinical measurement. Lancet, 1(8476), pp.307–310.
2) Giavarina, D., 2015. Understanding Bland Altman analysis. Biochemia Medica, 25(2), pp.141–151
}
