% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAanalysis.R
\name{BA.analysis}
\alias{BA.analysis}
\title{Bland-Altman analysis}
\usage{
BA.analysis(a, b, ...)
}
\arguments{
\item{a}{A numerical vector method A}

\item{b}{A numerical vector method B}

\item{percent}{bool for relative estimation (default F)}

\item{conf.int}{Confidence interval (default 0.95)}
}
\value{
List if the element required to a BA analysis
}
\description{
This function returns the Bland-Altman (BA) elements with 95% confidence
interval (by default) on the bias and on the limits of agreements. It also
calculates the percentage of error. Data can be converted to
relative estimation. Validation of differences normality with
Shapiro-Wilk test.
}
\details{
References:

1) Bland, J.M. & Altman, D.G., 1986. Statistical methods for assessing agreement
between two methods of clinical measurement. Lancet, 1(8476), pp.307–310.

2) Giavarina, D., 2015. Understanding Bland Altman analysis. Biochemia Medica, 25(2), pp.141–151
}
\examples{
ba <- BA.analysis(c(1,2,3,4), c(2,3,4,5))
}
