% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAplot.R
\name{BA.plot}
\alias{BA.plot}
\title{Bland-Altman plot}
\usage{
BA.plot(a, b, ...)
}
\arguments{
\item{x}{A numerical vector}

\item{y}{A numerical vector}

\item{title}{The title of the plot}

\item{percent}{bool for relative estimation (default F)}

\item{xlim}{c(x,y) limits on x axis}

\item{ylim}{c(x,y) limits on y axis}

\item{xlab}{x-axis title}

\item{ylab}{y-axis title}

\item{alpha}{point alpha (delfault 0.6)}

\item{pch}{point type (default 19)}

\item{conf.int}{Confidence interval (default 0.95)}

\item{hide.conf.int}{Hide confidence intervals (default FALSE)}

\item{reference}{c("A", "B", "mean") method to be used on X axis (default "mean")}

\item{save}{path to save plot (PDF). If only set to TRUE, title is used.}

\item{size}{c(width,height) in inch, default c(11,8.5)}
}
\value{
Base R plot
}
\description{
This function returns a Bland-Altman (BA) plot with 95% confidence
interval. It also calculates the percentage of error. Data can be converted to
relative estimation. Validation of differences normality with
Shapiro-Wilk test.
}
\examples{
BA.plot(c(1,2,3,4), c(2,3,4,5), title="My great title")
}
