% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAplot.R
\name{BA.plot}
\alias{BA.plot}
\title{Bland-Altman plot}
\usage{
BA.plot(a, b, ...)
}
\arguments{
\item{x}{A numerical vector}

\item{y}{A numerical vector}

\item{title}{The title of the plot}

\item{percent}{bool for relative estimation (default F)}

\item{conf.int}{Confidence interval (default 0.95)}

\item{reference}{c("A", "B", "mean") method to be used on X axis (default "mean")}
}
\value{
Standard R plot
}
\description{
This function returns a Bland-Altman (BA) plot with 95% confidence
interval (by default) on the bias and on the limits of agreements. It also
calculates the percentage of error. Data can be converted to
relative estimation. Validation of differences normality with
Shapiro-Wilk test.
}
\examples{
BA.analysis(c(1,2,3,4), c(2,3,4,5), title="My great title")
}
